raspberry/requirements.txt
cmake==3.25.2
wheel==0.40.0
dlib==19.24.1
numpy==1.22.4
click==8.1.3
pillow>=9.2.0
face-recognition==1.3.0

python-telegram-bot==20.3
paho-mqtt==1.6.1
requests==2.31.0